assembler: move error data and functions to new modules
