+++ assembler/src/util.rs
@@ -1,0 +1,17 @@
+pub(crate) fn min_signed_width(n: i32) -> u8 {
+    let mut width = 1;
+    const BASE: i32 = 2;
+    while n < -BASE.pow(width - 1) || n >= BASE.pow(width - 1) {
+        width += 1;
+    }
+    width as u8
+}
+
+pub(crate) fn min_unsigned_width(n: i32) -> u8 {
+    let mut width = 1;
+    const BASE: i32 = 2;
+    while n >= BASE.pow(width) {
+        width += 1;
+    }
+    width as u8
+}
